==> california-housing/notebooks/01-initial-data-prep.ipynb <==
# Initial data prep

This notebook takes the raw data (features and labels) and splits it into two subsets:

- Train (used for training)
- Test (used for final model evaluation)

> Two separate Assets will likely be created on Highwind, one for each subset of the data

# Config
RANDOM_SEED = 42
TEST_FRACTION = 0.2
SAVE_DIR = "../data/"
TARGET_COLUMN = "MedHouseVal"

import os
from sklearn.datasets import fetch_california_housing
from sklearn.model_selection import train_test_split
import pandas as pd

# Load raw data (features and labels)
california_housing = fetch_california_housing(as_frame=True)
df = california_housing.frame
df.head(3)

# Check null values
print(f"Total null values: {df.isna().sum().sum()}")

# Split features and labels
X = df.copy()
y = X.pop(TARGET_COLUMN)

# Check shapes
print(f"X: {X.shape}")
print(f"y: {y.shape}")

# Train-test split
X_train, X_test, y_train, y_test = train_test_split(
    X,
    y,
    test_size=TEST_FRACTION,
    random_state=RANDOM_SEED
)

# Recombine features and labels

# Train
df_train = X_train.copy()
df_train[TARGET_COLUMN] = y_train

# Test
df_test = X_test.copy()
df_test[TARGET_COLUMN] = y_test

# Check shapes
print(f"df_train: {df_train.shape}")
print(f"df_test: {df_test.shape}")

# Inspect
df_train.head(3)

# Save subsets separately
df_train.to_csv(os.path.join(SAVE_DIR, "train.csv"), index=False)
df_test.to_csv(os.path.join(SAVE_DIR, "test.csv"), index=False)
