==> california-housing/notebooks/03-model-inference.ipynb <==
# Model inference

This notebook shows how to load the trained model for inference. This includes processing new data to prepare it for inference.

> This process will likely be captured in a Docker Repo asset on Highwind

# Config
ARTIFACT_SAVE_DIR = "../saved_model/"

# For this example, the test data will be used for inference
TEST_DATA_PATH = "../data/test.csv"
TARGET_COLUMN = "MedHouseVal"

import os
import pandas as pd
import joblib
from huggingface_hub import hf_hub_download

## Load data

test_df = pd.read_csv(TEST_DATA_PATH)
test_df.head(3)

# Separate features and labels
X_test = test_df.copy()
y_test = X_test.pop(TARGET_COLUMN)

# Check shapes
print(f"X_test: {X_test.shape}")
print(f"y_test: {y_test.shape}")

## Download trined model and feature scaler

# Download from Hugging Face
download_files = ["model.joblib", "scaler.joblib"]

for file_name in download_files:
    hf_hub_download(
        repo_id="MelioAI/california-housing",
        filename=file_name,
        local_dir=ARTIFACT_SAVE_DIR
    )

## Process data

Repeat all processing steps to prepare data for inference

# Feature scaling
scaler = joblib.load(os.path.join(ARTIFACT_SAVE_DIR, "scaler.joblib"))
X_test_scaled = scaler.transform(X_test)

## Make predictions

# Load model
model = joblib.load(os.path.join(ARTIFACT_SAVE_DIR, "model.joblib"))
model

# Make predictions on processed data
y_pred = model.predict(X_test_scaled)
y_pred
